requirements.txt
python-nmap
pysimplegui